requirements.txt
python-telegram-bot==20.6
aiohttp==3.9.1
python-dotenv==1.0.0
base58==2.1.1
requests==2.31.0
beautifulsoup4==4.12.2
pymongo==4.6.1
httpx==0.25.1
playwright==1.41.1
solders==0.19.0
asyncio==3.4.3
python-dateutil==2.8.2
